STUFF ADDED!
As a customer, I can view testimonials.

As a customer, I can write testimonials and view them

As a customer, I can request a copy of the pamphlet, and if not on the
list add the needed information to be on the list. If already on the
list, a customer can request another copy of the pamphlet.

As an administrator we have to add all members who were not on the list,
but who gave the information to be on the list.

updates also to the output and how it is formatted

lots of file working going on.